varios

+++ src/components/exercisesList.js
@@ -2,13 +2,15 @@ import React, { Component } from 'react'
 import axios from 'axios';
 import Exercise from './exercise';
 import moment from 'moment'
+import { toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
 
 export default class ExercisesList extends Component {
     constructor(props) {
         super(props);
 
         this.deleteExercise = this.deleteExercise.bind(this);
-        //this.fetchExercises = this.fetchExercises.bind(this);
+        //console.log(this.props.history);
 
         this.state = {
             exercises: []
@@ -28,12 +30,18 @@ export default class ExercisesList extends Component {
     }
 
     deleteExercise(id) {
-        console.log('id: ' + id);
-
-        axios.delete('http://localhost:5000/exercises/' + id)
-            .then(res => console.log(res.data));
-
-        this.fetchExercises();
+        if (window.confirm('Are you sure you want to delete it?')) {
+            axios.delete('http://localhost:5000/exercises/' + id)
+                .then(res => { 
+                    toast.success(res.data.status);
+                })
+                .catch(err => { 
+                    console.log(err);
+                    toast.error(err.description);
+                });
+            
+                this.fetchExercises();
+        }
     }
 
     render() {
@@ -44,10 +52,11 @@ export default class ExercisesList extends Component {
                         this.state.exercises.map((item) => {
                             return (<Exercise
                                         key={item._id}
+                                        _id={item._id}
                                         description={item.description}
                                         username={item.username}
                                         date={moment(item.date).format('L')}
-                                        deleteExercise={() => this.deleteExercise(item._id)}
+                                        deleteExercise={this.deleteExercise}
                                     />)
                         })
                     }
